+++ CMakeLists.txt
@@ -281,9 +281,9 @@ add_executable(savegamedump WIN32
 	src/FileSystem.cpp
 	src/utils.cpp
 	src/StringF.cpp
-	src/GZipFormat.cpp
 	src/DateTime.cpp
 	src/Lang.cpp
+	src/core/GZipFormat.cpp
 	${FILESYSTEM_CXX_FILES}
 	# This one's in here because there's no such thing as separation in the codebase.
 	src/PngWriter.cpp
