+++ src/SpaceStationType.cpp
@@ -5,7 +5,6 @@
 #include "FileSystem.h"
 #include "Json.h"
 #include "MathUtil.h"
-#include "OS.h"
 #include "Pi.h"
 #include "Ship.h"
 #include "StringF.h"
