+++ src/SpeedLines.cpp
@@ -5,7 +5,7 @@
 
 #include "FileSystem.h"
 #include "Frame.h"
-#include "IniConfig.h"
+#include "core/IniConfig.h"
 #include "Pi.h"
 #include "Ship.h"
 #include "graphics/Renderer.h"
