+++ src/core/GZipFormat.cpp
