+++ src/core/GZipFormat.h
