+++ src/core/IniConfig.cpp
