+++ src/core/LZ4Format.cpp
