+++ src/core/LZ4Format.h
