+++ src/graphics/Graphics.cpp
@@ -4,7 +4,6 @@
 #include "Graphics.h"
 #include "FileSystem.h"
 #include "Material.h"
-#include "OS.h"
 #include "Renderer.h"
 #include "StringF.h"
 #include "utils.h"
