+++ src/graphics/opengl/RendererGL.cpp
@@ -3,7 +3,6 @@
 
 #include "RendererGL.h"
 #include "GLDebug.h"
-#include "OS.h"
 #include "Program.h"
 #include "RefCounted.h"
 #include "RenderStateGL.h"
