+++ src/posix/OSPosix.cpp
@@ -2,7 +2,7 @@
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "FileSystem.h"
-#include "OS.h"
+#include "core/OS.h"
 #include "buildopts.h"
 #include "utils.h"
