+++ src/savegamedump.cpp
@@ -2,7 +2,7 @@
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "FileSystem.h"
-#include "GZipFormat.h"
+#include "core/GZipFormat.h"
 #include "Json.h"
 #include <SDL.h>
