+++ src/scenegraph/BinaryConverter.cpp
@@ -3,10 +3,10 @@
 #include "BinaryConverter.h"
 #include "FileSystem.h"
 #include "GameSaveError.h"
-#include "LZ4Format.h"
 #include "NodeVisitor.h"
 #include "Parser.h"
 #include "StringF.h"
+#include "core/LZ4Format.h"
 #include "scenegraph/Animation.h"
 #include "scenegraph/Label3D.h"
 #include "scenegraph/MatrixTransform.h"
