+++ src/ship/ShipController.cpp
@@ -2,7 +2,7 @@
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "ShipController.h"
-#include "OS.h"
+#include "core/OS.h"
 #include "Ship.h"
 
 void ShipController::StaticUpdate(float timeStep)
