+++ src/uitest.cpp
@@ -2,9 +2,9 @@
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "FileSystem.h"
-#include "Lua.h"
-#include "OS.h"
-#include "PropertiedObject.h"
+#include "lua/Lua.h"
+#include "core/OS.h"
+#include "lua/PropertiedObject.h"
 #include "SDL.h"
 #include "graphics/Graphics.h"
 #include "graphics/Renderer.h"
